Collisions

+++ Handout2-Solution/Source/ModulePhysics.h
@@ -46,8 +46,8 @@ public:
 	bool CleanUp();
 
 	PhysBody* CreateCircle(int x, int y, int radius);
-	PhysBody* CreateRectangle(int x, int y, int width, int height, b2BodyType Type, int rotation);
-	PhysBody* CreateRectangleRebote(int x, int y, int width, int height, b2BodyType Type, int rotation);
+	PhysBody* CreateRectangle(int x, int y, int width, int height, b2BodyType Type, float rotation);
+	PhysBody* CreateRectangleRebote(int x, int y, int width, int height, b2BodyType Type, float rotation);
 	PhysBody* CreateRectangleSensor(int x, int y, int width, int height);
 	PhysBody* CreateChain(int x, int y, const int* points, int size);
